refactor: moved .key_generator and .date_generator to Inputs class
