+++ lib/enigma.rb
@@ -1,27 +1,28 @@
   require 'date'
+  require './inputs'
 
   class Enigma
 
     #1 helper for encrypt/decrypt
-    def key_generator(*details)   
-      if !details[1].nil? && details[1].size == 5 
-        key = details[1]
-      else
-        random_4_digits = rand.to_s[3..6]
-        key = "0#{random_4_digits.to_i}"
-      end
-    end
+    # def key_generator(*details)   
+    #   if !details[1].nil? && details[1].size == 5 
+    #     key = details[1]
+    #   else
+    #     random_4_digits = rand.to_s[3..6]
+    #     key = "0#{random_4_digits.to_i}"
+    #   end
+    # end
 
-    #2 helper for encrypt/decrypt
-    def date_generator(*details) 
-      if !details[1].nil? && details[1].size == 6
-       date = details[1] 
-      elsif !details[2].nil?
-        date = details[2]
-      elsif details[1].nil?
-          date = Date.today.strftime("%m%d%C")
-      end
-    end
+    # #2 helper for encrypt/decrypt
+    # def date_generator(*details) 
+    #   if !details[1].nil? && details[1].size == 6
+    #    date = details[1] 
+    #   elsif !details[2].nil?
+    #     date = details[2]
+    #   elsif details[1].nil?
+    #       date = Date.today.strftime("%m%d%C")
+    #   end
+    # end
 
     #3 helper for #5 - shifts_hash
     def keys_hash(key)
